refactor(rest-api-server): consolidate tRPC OpenAPI setup into a single function
