--- apps/rest-api-server/src/openapi.ts
@@ -1,15 +1,0 @@
-import { generateOpenApiDocument } from "trpc-openapi";
-
-import { env } from "@blobscan/env";
-
-import { appRouter } from "./app-router";
-
-// Generate OpenAPI schema document
-export const openApiDocument = generateOpenApiDocument(appRouter, {
-  title: "Blobscan API",
-  description: "OpenAPI compliant REST API built using tRPC with Express",
-  version: "1.0.0",
-  baseUrl: `${env.BLOBSCAN_API_BASE_URL}`,
-  docsUrl: "https://docs.blobscan.com/",
-  tags: ["blobs", "transactions", "blocks", "stats", "indexer", "system"],
-});
